+++ imports/ui/i18n/DateTimeLocalizationProvider.tsx
@@ -3,7 +3,8 @@ import React from 'react';
 import {LocalizationProvider} from '@mui/x-date-pickers/LocalizationProvider';
 import {AdapterDateFns} from '@mui/x-date-pickers/AdapterDateFns';
 
-import {useLocale, useLocaleText} from '../../i18n/datetime';
+import {useLocale} from '../../i18n/datetime';
+import {useLocaleText} from '../../i18n/datePickers';
 
 const DateTimeLocalizationProvider = ({children}) => {
 	const locale = useLocale();
